Use MVVM pattern
