+++ AroundTheWorld.xcodeproj/xcuserdata/giuseppe.sardo.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "089BC2CA-C2BC-4C1E-95BD-3CA7D0E44095"
+   type = "1"
+   version = "2.0">
+</Bucket>
